Examiner Part Added

+++ ExamSystem/Models/ResultDetail.cs
@@ -5,7 +5,7 @@
         public ResultDetail() { }
         public string ExamName { get; set; }
         public string ExamineeName { get; set; }   
-        public int ExamScore { get; set; }
+        public float ExamScore { get; set; }
         public int ScoreToPass { get; set; } 
         public string CorrectAnswers { get; set; }
         public String DateTaken { get; set; }
